Command line is now available and I fixed a problem i just discovered
The problem i discovered was that messages without a name associated
with them were not getting processed and it would show the phone number
twice when this happened. I set it so that it does get counted by making
the name "None" and i halved the phone number when the length was equal
to 20.